ml/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
grpcio==1.64.1
grpcio-tools==1.64.1
protobuf==5.27.1
torch==2.3.1
numpy==1.26.4
hypothesis==6.103.1
pytest==8.2.2
pytest-asyncio==0.23.7
